chore: separe in a few files. add a .envrc.
