+++ .envrc
@@ -1,0 +1,11 @@
+export NOSTR_PRIVATE_KEY=
+
+export LND1_TLS=
+export LND1_MACAROON=
+export LND1_DNS=
+
+export REDIS_PASSWORD=
+export REDIS_0_DNS=
+export REDIS_1_DNS=
+export REDIS_2_DNS=
+export REDIS_MASTER_NAME=
