+++ Dockerfile
@@ -2,7 +2,7 @@ FROM node:18-alpine AS BUILD_IMAGE
 
 WORKDIR /app
 
-RUN apk update && apk add git
+RUN apk update && apk add git && apk add python3 && apk add make && apk add g++
 
 COPY ./package.json ./yarn.lock ./
